We have added mandatory attributes for the XML tags  "maximum", "minimum" and "userid".
The "maximum" and "minimum" would contain the numeric bounds for a physical measurement, whether actual observed values or calibration curve ranges (i dont know).  In any event, the value contained could be "exclusive" or "included".  We are going to set to exclusive=false (or included). When you specify that the <maximum exclusive='false'>40</maximum>, we will assume that 40 is a possible value. In version 2.0, we may consider adding a tag to choose whether such value is possible or not.  It is all fine grain details that could be circumvented with practices (I.e; say 39.9999999999 if 40 is a rogue maximum).

The <userid> refers to a primary key in some system. The guidelines suggest a mandatory attribute that would specify an URL or some unreachable catalog (??!!??).  We just give the default website as the value of the attribute 
<userid directory="http://THISWEBSITE">123456</userid>